Move artifacts into their own internal package
