+++ internal/artifact/artifact_searcher_test.go
@@ -1,4 +1,4 @@
-package agent
+package artifact
 
 import (
 	"context"
