+++ internal/artifact/artifactory_downloader.go
@@ -1,4 +1,4 @@
-package agent
+package artifact
 
 import (
 	"context"
