+++ internal/artifact/artifactory_downloader_test.go
@@ -1,4 +1,4 @@
-package agent
+package artifact
 
 import (
 	"testing"
